django_api/requirements.txt
asgiref==3.7.2
Django==5.0.3
pip==22.3.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2024.1

wheel~=0.38.4
djangorestframework~=3.14.0